requirements.txt
biopython==1.79
cffi==1.14.5
greenlet==0.4.13
numpy==1.21.2
Pillow==8.4.0
readline==6.2.4.1
